requirements.txt
Django>=4.2,<5.0
django-environ>=0.11
psycopg[binary]>=3.1
djangorestframework>=3.14
drf-spectacular>=0.26
djangorestframework-simplejwt>=5.3
django-filter>=23.2
channels>=4.0
channels-redis>=4.1
celery>=5.3
redis>=5.0
django-guid>=3.4
